refactor: move proposals modals to sub-folder
