+++ frontend/svelte/src/lib/modals/proposals/ProposerModal.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-    import Modal from "./Modal.svelte";
+    import Modal from "../Modal.svelte";
     import type {ProposalInfo} from '@dfinity/nns';
     import {NeuronId} from '@dfinity/nns';
-    import {i18n} from "../stores/i18n";
-    import Card from "../components/ui/Card.svelte";
+    import {i18n} from "../../stores/i18n";
+    import Card from "../../components/ui/Card.svelte";
 
     export let proposalInfo: ProposalInfo;
