+++ frontend/svelte/src/tests/lib/modals/ProposalsFilterModal.spec.ts
@@ -7,7 +7,7 @@ import { fireEvent, render, waitFor } from "@testing-library/svelte";
 import { tick } from "svelte";
 import { get } from "svelte/store";
 import { DEFAULT_PROPOSALS_FILTERS } from "../../../lib/constants/proposals.constants";
-import ProposalsFilterModal from "../../../lib/modals/ProposalsFilterModal.svelte";
+import ProposalsFilterModal from "../../../lib/modals/proposals/ProposalsFilterModal.svelte";
 import { proposalsFiltersStore } from "../../../lib/stores/proposals.store";
 import type { ProposalsFilterModalProps } from "../../../lib/types/proposals";
 import { enumKeys } from "../../../lib/utils/enum.utils";
